Extract online communication test to separate class
